Adds client SetRetryCheck method.
This allows configuration of the retry logic to support non-standard
use cases like waiting for a URL to become available (retry on 404,
succeed on 200).

Instead of the RetryOnHTTP429 flag, the RetryCheck function can be
provided to decide whether to retry given the HTTP response.

SetRetryOnHTTP429 is kept for backwards compatibility but this is a
backwards incompatible change for users setting RetryOnHTTP429
directly on the client.